requirements.txt
Flask
python-dotenv
Flask-SQLAlchemy
python-telegram-bot
psycopg2-binary
celery
redis
